group database related internal apis
